clean up futurized corehq.apps.app_manager.views

+++ corehq/apps/app_manager/views/modules.py
@@ -765,8 +765,8 @@ def edit_module_detail_screens(request, domain, app_id, module_id):
         if case_list_lookup is not None:
             _save_case_list_lookup_params(detail.short, case_list_lookup, lang)
 
-    if int is not None:
-        detail.long.columns = list(map(DetailColumn.from_json, int))
+    if long is not None:
+        detail.long.columns = list(map(DetailColumn.from_json, long))
         if tabs is not None:
             detail.long.tabs = list(map(DetailTab.wrap, tabs))
     if filter != ():
